[QOL-7970] restore controller path
- move shared functions to new file and restore controller path so pages don't have to change
